intitial version: private browsing triggered when X-Private encountered.
General idea will now be: if there is at least one X-Private site, we switch
to/remain in private browsing. We may have to keep track of X-Private
sites. Challenge is when we have switched to private browsing: monitor requests
so that, when there is no remaing X-Private site (on all tabs and windows!),
we automatically switch back normal browsing.